[uci-defaults]: define wireless configuration
by default the dumbaps are not enabled because we do not want a roaming conflict with same SSID, this needs to be configured subtle.

also the effects are not known with multipsk and 802.11r, some suggest psk_generate_local needs to be turned off.

this however opens new options which I have yet to understand the workings of it, so as for now we temporary allow this 'bad configuration' until I notice the claim is valid.