[uci-defaults]: improve dumbap config
after alot of testing it seems we cannot use 192.168.1.1/24 atleast not for my full network.

First I thought it was a DSA issue since the eth1 is not bound to the switch cpu which lan1-lan5 does,
but this doesnt seem to be the case of the weird connection loss / time out.

this will allow us to combine eth1 or eth0 with lan, so we can cross vlan other AP.