[uci-defaults]: enable Generate PMK locally again.
Since we applied this fix for wifi roaming (802.11r) and multi psk environment turning off this checkbox is not suitable and causes adversed effects.
Even though, it should work as expected when everything is set to blank, yet this bugs.

The real fix to the firmware bug is setting the reassociation deadline to 20000 this way roaming and multi psk works fine, without redoing the 4-way handshake over and over again.

However with the local PMK checkbox unchecked, the unexpected result showed disconnections similar observed on a old qualcomm based TP-Link 1043nd router
which is in my notes a bad router with lots of driver quirk issues, while it being a Qualcomm device and this bug happens on a Mediatek based device, I strongly believe I hit another bug inside
the calibration firmware blob of Mediatek, perhaps it gives another developer some usefull insight why I think it are two bugs although Qualcomm has nothing to do with this.