[uci-defaults]: fix 106_set-pbr-wan-mark
we discovered that the default setting had no destination zone this is wrong.
instead we add a destination rather than towards the router, which would likely fail.

we also need to make sure leaking is not possible if PBR somehow fails and the mark for wan goes over wgclients route,
this is usually not possible, but if the package is doing silly things we have to add another extra step.
such problem can easily be caused when there is dns priority or a failing split tunnel.