[uci-defaults]: fix log levels, and some changes to vlan_naming in conjunction with tagged_vlan_interface
alot is unknown, and also ambigious, currently when wifi-station and wifi-vlan are involved,
luci automaticly assumes that these options are invalid, they are still a valid configuration instead
of being depended on wpa_psk_file as requirement, this is not required because wifi-station and wifi-vlan override the wpa_psk_file, it is basicly the same.

but now due this exact difference ambigious.